core/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agentsec-core"
version = "0.1.0"
description = "AgentSec core agent and skills library"
requires-python = ">=3.10"
dependencies = [
    "agent-framework-core==1.0.0b260107",
    "agent-framework-azure-ai==1.0.0b260107",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
